requirements.txt
apache-airflow-providers-postgres==5.8.0
pandas==2.1.4
pyarrow==16.1.0
requests==2.32.3
psycopg2-binary==2.9.9
pytest==8.2.0